+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 1.0.0-beta.3
 
 - `SequenceAction.addConnection` is now deprecated
+- publish to npm
 
 ### Bug fixes